+++ CodinGame/MarsLander/Models/Distance.cs
@@ -2,8 +2,8 @@ namespace CodinGame.MarsLander.Models
 {
     public class Distance
     {
-        public int VerticalDistance { get; set; }
-        public int HorizontalDistance { get; set; }
+        public double VerticalDistance { get; set; }
+        public double HorizontalDistance { get; set; }
         public double FullDistance { get; set; }
     }
 }
