+++ CodinGame/MarsLander/Models/Situation.cs
@@ -2,13 +2,13 @@ namespace CodinGame.MarsLander.Models
 {
     public class Situation
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
         public int Rotation { get; set; }
         public int Fuel { get; set; }
         public int Power { get; set; }
-        public int VerticalSpeed { get; set; }
-        public int HorizontalSpeed { get; set; }
+        public double VerticalSpeed { get; set; }
+        public double HorizontalSpeed { get; set; }
 
         public Situation Clone()
         {
